Stop using oversized types to encode unsigned numeric values and use unsigned methods instead; Implement binary encoding of types

+++ src/main/java/com/opencastsoftware/wasm4j/Func.java
@@ -6,17 +6,17 @@ import com.opencastsoftware.wasm4j.types.ValType;
 import java.util.List;
 
 public class Func {
-    private final long typeIndex;
+    private final int typeIndex;
     private final List<ValType> locals;
     private final List<Instruction> body;
 
-    public Func(long typeIndex, List<ValType> locals, List<Instruction> body) {
+    public Func(int typeIndex, List<ValType> locals, List<Instruction> body) {
         this.typeIndex = typeIndex;
         this.locals = locals;
         this.body = body;
     }
 
-    public long typeIndex() {
+    public int typeIndex() {
         return typeIndex;
     }
